Render a dupe of the post to avoid double rendering
I noticed I was having problems caused by my posts being rendered twice.
When a post is rendered, the result is stored and will be used as input
to the renderer if you call render() again. Since rssgenerator.rb
renders a post and then it gets rendered again when generating the
entire site, the post gets double rendered. This causes problems if the
post contains {% raw %}, as the raw content will get processed the
second time around.